Favourites Post started working
when u click to add a post as favorurite, the post id and user id is added to the favourite posts table. All we need to do now is display them on the favourite posts page and after I merge to main, I need to make sure that the image upploading is working iwht the posts function.